+++ index.js
@@ -1,5 +1,5 @@
 var next = require('./lib/index.js');
 
-next.version = '1.27.13';
+next.version = '1.27.15';
 
 module.exports = next;