+++ CHANGELOG.md
@@ -1,5 +1,43 @@
 # Change Log 
 
+## [1.27.15](https://github.com/alibaba-fusion/next/compare/1.27.14...1.27.15) (2024-07-05)
+
+
+### Features
+
+* **Tree:** the Tree component provides the scrollFilterNodeIntoView API ([#4860](https://github.com/alibaba-fusion/next/issues/4860)) ([9df35fe](https://github.com/alibaba-fusion/next/commit/9df35fea5efdd7b2b87926f308eb7e4f7eefafd0))
+
+
+### Bug Fixes
+
+* **DatePicker2:** unable to enter space to enter time ([65faba2](https://github.com/alibaba-fusion/next/commit/65faba254b349ead3642b3efafab67023ed98a07))
+
+
+## [1.27.14](https://github.com/alibaba-fusion/next/compare/1.27.13...1.27.14) (2024-07-03)
+
+
+### Features
+
+* **Balloon:** add mouseEnterDelay and mouseLeaveDelay, with higher priority than delay ([1f6ccdd](https://github.com/alibaba-fusion/next/commit/1f6ccdd7a3aa7337946055b1a41b5889cdc82f60))
+* **CascaderSelect:** clear the search box after selecting the item, close [#3008](https://github.com/alibaba-fusion/next/issues/3008) [#3415](https://github.com/alibaba-fusion/next/issues/3415) ([9e159f2](https://github.com/alibaba-fusion/next/commit/9e159f26ddb0c4e575be00e503529c5d6da74fde))
+
+
+### Bug Fixes
+
+* **CascaderSelect:** use white background with virtual scrolling ([65a206b](https://github.com/alibaba-fusion/next/commit/65a206bc79a191c9c4bf72388861c6ebf629b11a))
+* **DatePicker2:** okButton is still clickable when input value is in disabled date, close [#3801](https://github.com/alibaba-fusion/next/issues/3801) ([#4870](https://github.com/alibaba-fusion/next/issues/4870)) ([df4ecc7](https://github.com/alibaba-fusion/next/commit/df4ecc742b0f0646a69f2387980c0dc655c7ec5e))
+* **DatePicker2:** unable to enter space to enter time ([83269e0](https://github.com/alibaba-fusion/next/commit/83269e0ad3c02c921072ab16470d46a19bcdf9f0))
+* **Message:** the styles of the outer Message component will affect the inner component when nested, close [#4851](https://github.com/alibaba-fusion/next/issues/4851) ([ec8978e](https://github.com/alibaba-fusion/next/commit/ec8978e223078fb32d45d2c499bd59d07dabe63d))
+* **Select:** should be able to not clear search value in multiple mode, close [#3590](https://github.com/alibaba-fusion/next/issues/3590) ([ab017bc](https://github.com/alibaba-fusion/next/commit/ab017bc1358ca7cee2b1dfba92ecda3a4dbca9dd))
+* **Slider:** the last element covers on top when the Slider component is set to fade effect ([28a6b23](https://github.com/alibaba-fusion/next/commit/28a6b235c10790f676f33bf221f4623c7c62bc3f))
+* **Tree:** should keep click and key events consistent,close [#2899](https://github.com/alibaba-fusion/next/issues/2899) ([906bc90](https://github.com/alibaba-fusion/next/commit/906bc90893fecfb0971ee2e0fc4b7be324a469b8))
+
+
+### Documentation
+
+* update contributing document ([740d523](https://github.com/alibaba-fusion/next/commit/740d523f6d9c7115c66c498857f95b4fd2a2c781))
+
+
 ## [1.27.13](https://github.com/alibaba-fusion/next/compare/1.27.12...1.27.13) (2024-06-11)